+++ toroCLibrary/Library/mbctype_h/_setmbcp.c
@@ -25,7 +25,7 @@ Author:
 //
 // single byte codepages known to Microsoft C VS2022
 //
-extern short _cdeMsftSingleByteCodePageNumber[];
+extern unsigned short _cdeMsftSingleByteCodePageNumber[];
 
 /**
 
@@ -52,11 +52,11 @@ int _setmbcp(int codepage) {
 
     do {
 
-        for (i = 0; -1 != _cdeMsftSingleByteCodePageNumber[i]; i++)
-            if ((short)codepage == _cdeMsftSingleByteCodePageNumber[i])
+        for (i = 0; (unsigned short) -1 != _cdeMsftSingleByteCodePageNumber[i]; i++)
+            if ((unsigned short)codepage == _cdeMsftSingleByteCodePageNumber[i])
                 break;//for
 
-        if (-1 != _cdeMsftSingleByteCodePageNumber[i])
+        if ((unsigned short)-1 != _cdeMsftSingleByteCodePageNumber[i])
             break;
 
         //
