+++ toroCLibrary/Library/stdio_h/Fputs.c
@@ -49,5 +49,5 @@ NOTE:   Seems, that Microsoft has not implemented any errorreporting.
 **/
 int fputs(const char* s, FILE* stream) {
 
-    return EOF != fwrite(s, 1, strlen(s), stream) ? 0 : EOF;
+    return (size_t)EOF != fwrite(s, 1, strlen(s), stream) ? 0 : EOF;
 }
